requirements.txt
langchain
langchain-community
langchain-pinecone
langchain-cohere
pinecone-client
python-dotenv